pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.0.1"
description = "FastAPI Boilerplate"
authors = ["Foundation4Tech TCS"]
package-mode = false

[tool.poetry.dependencies]
aiomysql = "^0.2.0"
aiosqlite = "^0.20.0"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
bcrypt = "^4.2.1"
black = "^24.10.0"
click = "^8.1.7"
celery = "^5.4.0"
cryptography = "^44.0.0"
email-validator = "^2.2.0"
faker = "^33.1.0"
fastapi = "^0.115.0"
fastapi-event = "^0.1.3"
httpx = "^0.23.3"
greenlet = "^3.1.1"
gunicorn = "^23.0.0"
isort = "^5.13.2"
packaging = "^24.2"
passlib = "^1.7.4"
psycopg2 = "^2.9.10"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.0"
PyJWT = "^2.10.0"
pylint = "^3.3.2"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
python = "^3.11"
python-jose = "^3.3.0"
pyyaml = "^6.0.2"
redis = "^5.2.1"
SQLAlchemy = "^2.0.36"
ujson = "^5.10.0"
uvicorn = "^0.34.0"
wrapt = "^1.17.0"

[tool.poetry.dev-dependencies]
behave = "^1.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
